build(typescript): separate tsc config for building
For type checking we want to check everything (including tests), but for
building we only want to build src; unfortunately, tsc doesn't support
this configuration when you also want lint-staged to pass individual
files to tsc and eslint to use the type enhanced checking.

This fixes it by forking the tsconfig.json for each scenario.