fix(api): users can be queried by their sub
The whole username vs sub vs email vs phone number thing is confusing,
and in an earlier change when I allowed usernames to be supplied by the
client it broke the ability to lookup by the sub when you use an email.

This is made more confusing by how the userName parameter supplied to
various calls to Cognito can be either the sub or the username or one of
the other username attributes 🤷

This change allows you to _always_ be able to lookup a user by their sub
regardless of what username they've registered with.