refactor(db): store dates as ISO strings and parse
Dates were stored as numbers and the types reflected that, but the AWS
response types expect Dates so we were having to convert the types
before returning responses.

This change makes it so all our types are Dates, but any field named
*Date (which matches the AWS convention) will be serialized to an
ISODateString in storage and deserialized to a Date.

Any historical data stored as a number will be converted to a Date on
read, and if written back will be saved as an ISODateString.

This change should not have any impact on users.