fix: sub as username when using emails
If a User Pool is configured with email as a UsernameAttribute then:

1. The Username field in requests to create users should be validated as
   an email address
2. The User's `sub` field is used as their username attribute and is
   returned from any get or list actions

If a User Pool doesn't have a UsernameAttribute set, then the Username
from the create actions will be used as the User's Username.

BREAKING CHANGE: It's possible that users may have been stored with
their email address as their Username in the cognito-local database;
this was incorrect and shouldn't have happened. You may need to recreate
any users who were incorrectly saved with their email address as their
username.